Made the entirety of the image cards clickable in the database page
You can click anywhere on the image card now to go to the respective card's page now instead of just on the button. The mouse cursor was changed to a pointer to reflect this.
(Note: I did not remove the old button I just added an additional href property to the encapsulating container)